refactor code

+++ src/app/modules/result/result.component.ts
@@ -10,10 +10,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { debounce, filter } from 'lodash-es';
 import { Result } from '../../shared/models/result.model';
 import { ConfirmDialogComponent } from '../../shared/dialogs/confirm-dialog/confirm-dialog.component';
-import { ListeningComponent } from '../../tabs/listening/listening.component';
-import { MultipleChoicesComponent } from '../question/multiple-choices/multiple-choices.component';
 import { QuizService } from '../quizzes/quizzes.service';
-import { ShortAnswerComponent } from '../question/short-answer/short-answer.component';
 import { BandScorePipe } from './band-score.pipe';
 import { ResultService } from './result.service';
 import { InputPasswordDialogComponent } from '../../shared/dialogs/input-password-dialog/input-password-dialog.component';
@@ -22,13 +19,10 @@ import { InputPasswordDialogComponent } from '../../shared/dialogs/input-passwor
   selector: 'app-result',
   standalone: true,
   imports: [
-    MultipleChoicesComponent,
-    ShortAnswerComponent,
     FormsModule,
     MatCardModule,
     MatButtonModule,
     MatInputModule,
-    ListeningComponent,
     MatMenuModule,
     MatIcon,
     BandScorePipe,
@@ -61,23 +55,25 @@ export class ResultComponent {
       });
   }
 
-  view(id: string) {
+  view(id?: string) {
     const dialogRef = this.dialog.open(InputPasswordDialogComponent);
     dialogRef.afterClosed().subscribe((isMatch) => {
-      if (isMatch) {
+      if (isMatch && id) {
         this.router.navigate([`result-detail`, id]);
       }
     });
   }
 
-  continue(id: string) {
+  continue(id?: string) {
     const test = {
       testId: id,
     };
-    this.router.navigate(['continue-test', id], { state: test });
+    if (id) {
+      this.router.navigate(['continue-test', id], { state: test });
+    }
   }
 
-  onDeleteResultClick(resultId: string) {
+  onDeleteResultClick(resultId?: string) {
     const dialogRef = this.dialog.open(ConfirmDialogComponent, {
       hasBackdrop: true,
       disableClose: true,
@@ -86,7 +82,7 @@ export class ResultComponent {
     dialogRef.componentInstance.isWarning = true;
     dialogRef.componentInstance.message = 'Confirm to delete this?';
     dialogRef.afterClosed().subscribe((isConfirm: boolean) => {
-      if (isConfirm) {
+      if (isConfirm && resultId) {
         this.deleteResult(resultId);
       }
     });
@@ -94,7 +90,7 @@ export class ResultComponent {
 
   deleteResult(resultId: string) {
     this.resultService.deleteById(resultId).subscribe(() => {
-      this.results = filter(this.results, (result) => result.id !== resultId);
+      this.results = filter(this.results, (result) => result._id !== resultId);
     });
   }
